obscure private device provenance
Allow users to choose whether or not to reveal the provenance of the data to
consumers.  This is intended to help protect data rights to portability and
access.  If consumers start abusing or discriminating against Nightscout users
due to the source of the data, this allows users to conceal this information.