rm package-lock.json && npm install
npm ls fsevent didn't list any output.  I'm running on linux,
and regenerating this file after removing it seems to remove
the fsevent package from the file.
It's likely that fsevent may be used in development mode as part
of one of the build utilities, but it shouldn't block anyone from
deploying.

https://github.com/nightscout/cgm-remote-monitor/issues/7475